+++ services/dynamic-scheduler/tests/unit/api_rest/conftest.py
@@ -10,19 +10,6 @@ from pytest_simcore.helpers.typing_env import EnvVarsDict
 
 
 @pytest.fixture
-def app_environment(
-    disable_rabbitmq_setup: None,
-    disable_redis_setup: None,
-    disable_service_tracker_setup: None,
-    disable_deferred_manager_setup: None,
-    disable_notifier_setup: None,
-    disable_status_monitor_setup: None,
-    app_environment: EnvVarsDict,
-) -> EnvVarsDict:
-    return app_environment
-
-
-@pytest.fixture
 async def client(
     app_environment: EnvVarsDict, app: FastAPI
 ) -> AsyncIterator[AsyncClient]:
