+++ services/dynamic-scheduler/tests/unit/api_rest/test_api_rest__meta.py
@@ -1,12 +1,28 @@
+import pytest
+
 # pylint:disable=redefined-outer-name
 # pylint:disable=unused-argument
 from fastapi import status
 from httpx import AsyncClient
+from pytest_simcore.helpers.typing_env import EnvVarsDict
 from simcore_service_dynamic_scheduler._meta import API_VTAG
 from simcore_service_dynamic_scheduler.models.schemas.meta import Meta
 
 
-async def test_health(client: AsyncClient):
+@pytest.fixture
+def app_environment(
+    disable_rabbitmq_setup: None,
+    disable_redis_setup: None,
+    disable_service_tracker_setup: None,
+    disable_deferred_manager_setup: None,
+    disable_notifier_setup: None,
+    disable_status_monitor_setup: None,
+    app_environment: EnvVarsDict,
+) -> EnvVarsDict:
+    return app_environment
+
+
+async def test_meta(client: AsyncClient):
     response = await client.get(f"/{API_VTAG}/meta")
     assert response.status_code == status.HTTP_200_OK
     assert Meta.parse_raw(response.text)
